wallet mod describe
